# zenzefi_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZenZefi REST API provides operations for resources certificate, configuration, log, user, system",
        "version": "1.0",
        "title": "ZenZefi REST API"
    },
    "host": "localhost:61000",
    "basePath": "/",
...
    "paths": {
...
        "/users/currentUser": {
            "get": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Checks if the current user is the default user. Endpoint used in the ZenZefi UI for automated logout.",
                "operationId": "currentUserUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Logs in a user in based on the provided data (UC53).",
                "operationId": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "description": "The user.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success. User logged in successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "406": {
                        "description": "Login Failed - Unknown Username or wrong combination of Username and Password"
                    },
                    "412": {
                        "description": "A User is already logged in. Please logoff first. Login rejected."
                    },
                    "500": {
                        "description": "Exception message"
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Log out a user from ZenZefi.",
                "operationId": "logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/users/login": {
            "post": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi User Management"
                ],
                "summary": "Logs in a user in based on the provided data (UC53).",
                "operationId": "loginV1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "description": "The user.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success. User logged in successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "406": {
                        "description": "Login Failed - Unknown Username or wrong combination of Username and Password"
                    },
                    "412": {
                        "description": "A User is already logged in. Please logoff first. Login rejected."
                    },
                    "500": {
                        "description": "Exception message"
                    }
                }
            }
        },
        "/v1/users/logout": {
            "post": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi User Management"
                ],
                "summary": "Log out a user from ZenZefi.",
                "operationId": "logoutV1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "required": [
                "firstName",
                "lastName",
                "organisation",
                "userName",
                "userPassword"
            ],
            "properties": {
                "createTimestamp": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "description": "The first name. Length between 1 and 100 characters."
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string",
                    "description": "The last name. Length between 1 and 100 characters."
                },
                "organisation": {
                    "type": "string",
                    "description": "The organisation. Length between 1 and 100 characters."
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "DEFAULT",
                        "REGULAR"
                    ]
                },
                "userName": {
                    "type": "string",
                    "description": "The user name. Length between 1 and 7 characters. Must be alpha numeric, non blank."
                },
                "userPassword": {
                    "type": "string",
                    "description": "The user password. Length between 9 and 100 characters. Must contain upper case, lower case, digit, and special characters."
                }
            }
        },
...
    }
}